requirements.txt
asgiref==3.7.0
beautifulsoup4==4.12.2
boto3==1.26.163
botocore==1.29.163
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-bootstrap4==23.1
django-cloudinary-storage==0.3.0
django-environ==0.10.0
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
Pillow==9.5.0
psycopg2==2.9.6
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
Unipath==1.1
urllib3==1.26.16
whitenoise==6.4.0
